# swagger.json
{"openapi":"3.0.1","info":{"title":"Swagger","version":"1.0.1","description":"example","termsOfService":"http://example.com/terms/","contact":{"name":"API Support","url":"http://www.example.com/support","email":"support@example.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0.html","email":"example@gmail.com"}},"tags":[{"name":"person","description":"person operations"},{"name":"login","description":"login operations"}],"paths":{"/login":{"post":{"summary":"login","description":"Login to the application","tags":["login"],"produces":["application/json"],"parameters":[{"name":"username","description":"Username to use for login.","in":"formData","required":true,"type":"string"},{"name":"password","description":"User's password.","in":"formData","required":true,"type":"string"}],"responses":{"200":{"description":"login"}}}},"/person":{"post":{"description":"get person data what you input","tags":["person"],"summary":"get person","produces":["application/json"],"requestBody":{"requeired":true,"content":{"application/x-www-form-urlencoded":{"schema":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string","enum":["Male","Female","other"],"example":"Male"}},"required":["name"]}},"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string","enum":["Male","Female","other"],"example":"Male"}}},"examples":{"Mengty":{"value":{"name":"Mengty","gender":"Male"}},"Lakayata":{"value":{"name":"Lakayata","gender":"other"}}}}}},"responses":{"200":{"description":"get person data","content":{"application/json":{"schema":{"type":"object"},"example":{"name":"Mengty","gender":"Male"}}}}}},"put":{"description":"get person data what you input","tags":["person"],"summary":"get person","produces":["application/json"],"requestBody":{"requeired":true,"content":{"application/x-www-form-urlencoded":{"schema":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string","enum":["Male","Female","other"],"example":"Male"}}}},"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string","enum":["Male","Female","other"],"example":"Male"}}},"examples":{"Mengty":{"value":{"name":"Mengty","gender":"Male"}},"Lakayata":{"value":{"name":"Lakayata","gender":"other"}}}}}},"responses":{"200":{"description":"get person data","content":{"application/json":{"schema":{"type":"object"},"example":{"name":"Mengty","gender":"Male"}}}}}}},"/person/{id}":{"get":{"description":"get person data","tags":["person"],"summary":"get person","parameters":[{"in":"header","name":"x-custom","schema":{"type":"string"},"required":true},{"in":"path","name":"id","schema":{"type":"integer"},"required":true,"description":"Numeric id of person to get"},{"in":"query","name":"offset","schema":{"type":"integer"},"description":"offset data"},{"in":"query","name":"limit","schema":{"type":"integer"},"description":"limit data"}],"responses":{"200":{"description":"get person data","content":{"application/json":{"schema":{"type":"object"},"example":{"name":"Mengty","gender":"Male"}}}}}}}},"components":{}}